+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# Unreleased
+# 2.3.2
+* Ensure additional fields are reset after logging
+* Let Oj fallback to :object mode for non-representable data
+* Fix SimpleLogger to handle non-String log messages
 * Run RuboCop checks on build and fix RuboCop offenses
 
 # 2.3.1